add primitive func and more test
