+++ internal/daemon/controller/handlers/groups/grants_test.go
@@ -1,119 +1,426 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: BUSL-1.1
-
 package groups_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
+	"github.com/google/uuid"
 	"github.com/hashicorp/boundary/globals"
+	"github.com/hashicorp/boundary/internal/auth/password"
 	"github.com/hashicorp/boundary/internal/authtoken"
 	"github.com/hashicorp/boundary/internal/daemon/controller/auth"
 	"github.com/hashicorp/boundary/internal/daemon/controller/handlers/groups"
 	"github.com/hashicorp/boundary/internal/db"
 	pbs "github.com/hashicorp/boundary/internal/gen/controller/api/services"
+	authpb "github.com/hashicorp/boundary/internal/gen/controller/auth"
 	"github.com/hashicorp/boundary/internal/iam"
 	"github.com/hashicorp/boundary/internal/kms"
+	"github.com/hashicorp/boundary/internal/requests"
+	"github.com/hashicorp/boundary/internal/server"
 	"github.com/stretchr/testify/require"
 )
 
-// TestGrants_ReadActions tests read actions to assert that grants are being applied properly
-//
-//	 Role - which scope the role is created in
-//			- global level
-//			- org level
-//			- project level
-//		Grant - what IAM grant scope is set for the permission
-//			- global: descendant
-//			- org: children
-//			- project
-//		Scopes [resource]:
-//			- global [globalGroup]
-//				- org1 [org1Group]
-//					- proj1 [proj1Group]
-//				- org2 [org2Group]
-//					- proj2 [proj2Group]
-//					- proj3 [proj3Group]
-func TestGrants_ReadActions(t *testing.T) {
+// Test Dimension
+//  Role - which scope the role is created in
+// 		- global level
+// 		- org level
+// 		- project level
+//	Grant - what IAM grant scope is set for the permission
+// 		- global: descendant
+//		- org: children
+// 		- project
+//	Resource - where resources are created (group)
+// 		- global
+//			- org1
+// 				- project1
+
+func TestGrants_Get(t *testing.T) {
 	ctx := context.Background()
 	conn, _ := db.TestSetup(t, "postgres")
 	wrap := db.TestWrapper(t)
+	kmsCache := kms.TestKms(t, conn, wrap)
+	rw := db.New(conn)
 	iamRepo := iam.TestRepo(t, conn, wrap)
 	repoFn := func() (*iam.Repository, error) {
 		return iamRepo, nil
 	}
-	kmsCache := kms.TestKms(t, conn, wrap)
 	s, err := groups.NewService(ctx, repoFn, 1000)
 	require.NoError(t, err)
-	org1, _ := iam.TestScopes(t, iamRepo)
-	org2, proj2 := iam.TestScopes(t, iamRepo)
-	proj3 := iam.TestProject(t, iamRepo, org2.PublicId)
+
+	atRepo, err := authtoken.NewRepository(ctx, rw, rw, kmsCache)
+	require.NoError(t, err)
+	atRepoFn := func() (*authtoken.Repository, error) {
+		return atRepo, nil
+	}
+	serversRepoFn := func() (*server.Repository, error) {
+		return server.NewRepository(ctx, rw, rw, kmsCache)
+	}
+
+	org, proj := iam.TestScopes(t, iamRepo)
 	globalGroup := iam.TestGroup(t, conn, globals.GlobalPrefix, iam.WithDescription("global"), iam.WithName("global"))
-	org1Group := iam.TestGroup(t, conn, org1.GetPublicId(), iam.WithDescription("org1"), iam.WithName("org1"))
-	org2Group := iam.TestGroup(t, conn, org2.GetPublicId(), iam.WithDescription("org2"), iam.WithName("org2"))
+	orgGroup := iam.TestGroup(t, conn, org.GetPublicId(), iam.WithDescription("org"), iam.WithName("org"))
+	projGroup := iam.TestGroup(t, conn, proj.GetPublicId(), iam.WithDescription("project"), iam.WithName("project"))
 
-	proj2Group := iam.TestGroup(t, conn, proj2.GetPublicId(), iam.WithDescription("proj2"), iam.WithName("proj2"))
-	proj3Group := iam.TestGroup(t, conn, proj3.GetPublicId(), iam.WithDescription("proj3"), iam.WithName("proj3"))
+	authMethod := password.TestAuthMethods(t, conn, globals.GlobalPrefix, 1)[0]
+	testcases := []struct {
+		name      string
+		roleScope string
 
-	t.Run("List", func(t *testing.T) {
-		testcases := []struct {
-			name          string
-			input         *pbs.ListGroupsRequest
-			rolesToCreate []authtoken.TestRoleGrantsForToken
-			wantErr       error
-			wantIDs       []string
-		}{
-			{
-				name: "global role grant this and children returns global and org groups",
-				input: &pbs.ListGroupsRequest{
-					ScopeId:   globals.GlobalPrefix,
-					Recursive: true,
-				},
-				rolesToCreate: []authtoken.TestRoleGrantsForToken{
-					{
-						RoleScopeID:  globals.GlobalPrefix,
-						GrantStrings: []string{"ids=*;type=group;actions=list,read"},
-						GrantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeChildren},
-					},
-				},
-				wantErr: nil,
-				wantIDs: []string{globalGroup.PublicId, org1Group.PublicId, org2Group.PublicId},
+		grantStrings []string
+		grantScopes  []string
+		getIdFound   map[string]bool
+	}{
+		{
+			name:         "global_role_grant_this",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeThis},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: true,
+				orgGroup.PublicId:    false,
+				projGroup.PublicId:   false,
 			},
-			{
-				name: "org role grant this and children returns org and project groups",
-				input: &pbs.ListGroupsRequest{
-					ScopeId:   org2.PublicId,
-					Recursive: true,
-				},
-				rolesToCreate: []authtoken.TestRoleGrantsForToken{
-					{
-						RoleScopeID:  org2.PublicId,
-						GrantStrings: []string{"ids=*;type=group;actions=list,read"},
-						GrantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeChildren},
-					},
-				},
-				wantErr: nil,
-				wantIDs: []string{org2Group.PublicId, proj2Group.PublicId, proj3Group.PublicId},
+		},
+		{
+			name:         "global_role_grant_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
 			},
-		}
+		},
+		{
+			name:         "global_role_grant_descendant",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeDescendants},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "global_role_grant_this_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: true,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name:         "global_role_grant_this_descendant",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeDescendants},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: true,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "org_role_grant_this",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    org.GetPublicId(),
+			grantScopes:  []string{globals.GrantScopeThis},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name:         "org_role_grant_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    org.GetPublicId(),
+			grantScopes:  []string{globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    false,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "org_role_grant_this_and_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    org.GetPublicId(),
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name: "project_role_grant_this",
 
-		for _, tc := range testcases {
-			t.Run(tc.name, func(t *testing.T) {
-				tok := authtoken.TestAuthTokenWithRoles(t, conn, kmsCache, globals.GlobalPrefix, tc.rolesToCreate)
-				fullGrantAuthCtx := auth.TestAuthContextFromToken(t, conn, wrap, tok, iamRepo)
-				got, finalErr := s.ListGroups(fullGrantAuthCtx, tc.input)
-				if tc.wantErr != nil {
-					require.ErrorIs(t, finalErr, tc.wantErr)
-					return
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			grantScopes:  []string{globals.GrantScopeThis},
+			roleScope:    proj.GetPublicId(),
+
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    false,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "global_role_grant_all_scopes_specific_group_id",
+			grantStrings: []string{fmt.Sprintf("ids=%s;types=group ;actions=*", orgGroup.PublicId)},
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeDescendants},
+			roleScope:    globals.GlobalPrefix,
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name: "global_role_grant_all_specific_permissions",
+			grantStrings: []string{
+				fmt.Sprintf("ids=%s;types=group;actions=read", orgGroup.PublicId),
+				fmt.Sprintf("ids=%s;types=group;actions=read", projGroup.PublicId)},
+			grantScopes: []string{org.PublicId, proj.PublicId},
+			roleScope:   globals.GlobalPrefix,
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+	}
+
+	for _, tc := range testcases {
+		t.Run(tc.name, func(t *testing.T) {
+			// this creates everything required to get a token and creates context with auth token
+			acct := password.TestAccount(t, conn, authMethod.GetPublicId(), uuid.NewString())
+			user := iam.TestUser(t, iamRepo, globals.GlobalPrefix, iam.WithAccountIds(acct.GetPublicId()))
+			role := iam.TestRoleWithGrants(t, conn, tc.roleScope, tc.grantScopes, tc.grantStrings)
+			_ = iam.TestUserRole(t, conn, role.PublicId, user.PublicId)
+			fullGrantToken, err := atRepo.CreateAuthToken(ctx, user, acct.GetPublicId())
+			require.NoError(t, err)
+			fullGrantAuthCtx := auth.NewVerifierContext(requests.NewRequestContext(ctx, requests.WithUserId(user.GetPublicId())),
+				repoFn, atRepoFn, serversRepoFn, kmsCache, &authpb.RequestInfo{
+					PublicId:    fullGrantToken.PublicId,
+					Token:       fullGrantToken.GetToken(),
+					TokenFormat: uint32(auth.AuthTokenTypeBearer),
+				})
+			for id, found := range tc.getIdFound {
+				_, err := s.GetGroup(fullGrantAuthCtx, &pbs.GetGroupRequest{
+					Id: id,
+				})
+				// not found means expect error
+				if !found {
+					require.Error(t, err)
+					continue
 				}
-				require.NoError(t, finalErr)
-				var gotIDs []string
-				for _, g := range got.Items {
-					gotIDs = append(gotIDs, g.GetId())
+				require.NoError(t, err)
+
+			}
+		})
+	}
+
+}
+
+func TestGrants_List(t *testing.T) {
+	ctx := context.Background()
+	conn, _ := db.TestSetup(t, "postgres")
+	wrap := db.TestWrapper(t)
+	kmsCache := kms.TestKms(t, conn, wrap)
+	rw := db.New(conn)
+	iamRepo := iam.TestRepo(t, conn, wrap)
+	repoFn := func() (*iam.Repository, error) {
+		return iamRepo, nil
+	}
+	s, err := groups.NewService(ctx, repoFn, 1000)
+	require.NoError(t, err)
+
+	atRepo, err := authtoken.NewRepository(ctx, rw, rw, kmsCache)
+	require.NoError(t, err)
+	atRepoFn := func() (*authtoken.Repository, error) {
+		return atRepo, nil
+	}
+	serversRepoFn := func() (*server.Repository, error) {
+		return server.NewRepository(ctx, rw, rw, kmsCache)
+	}
+
+	org, proj := iam.TestScopes(t, iamRepo)
+	globalGroup := iam.TestGroup(t, conn, globals.GlobalPrefix, iam.WithDescription("global"), iam.WithName("global"))
+	orgGroup := iam.TestGroup(t, conn, org.GetPublicId(), iam.WithDescription("org"), iam.WithName("org"))
+	projGroup := iam.TestGroup(t, conn, proj.GetPublicId(), iam.WithDescription("project"), iam.WithName("project"))
+
+	authMethod := password.TestAuthMethods(t, conn, globals.GlobalPrefix, 1)[0]
+	testcases := []struct {
+		name      string
+		roleScope string
+
+		grantStrings []string
+		grantScopes  []string
+		getIdFound   map[string]bool
+	}{
+		{
+			name:         "global_role_grant_this",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeThis},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: true,
+				orgGroup.PublicId:    false,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name:         "global_role_grant_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name:         "global_role_grant_descendant",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeDescendants},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "global_role_grant_this_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: true,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name:         "global_role_grant_this_descendant",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    globals.GlobalPrefix,
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeDescendants},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: true,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "org_role_grant_this",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    org.GetPublicId(),
+			grantScopes:  []string{globals.GrantScopeThis},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name:         "org_role_grant_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    org.GetPublicId(),
+			grantScopes:  []string{globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    false,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "org_role_grant_this_and_children",
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			roleScope:    org.GetPublicId(),
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeChildren},
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name: "project_role_grant_this",
+
+			grantStrings: []string{"id=*;type=*;actions=*"},
+			grantScopes:  []string{globals.GrantScopeThis},
+			roleScope:    proj.GetPublicId(),
+
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    false,
+				projGroup.PublicId:   true,
+			},
+		},
+		{
+			name:         "global_role_grant_all_scopes_specific_group_id",
+			grantStrings: []string{fmt.Sprintf("ids=%s;types=group ;actions=*", orgGroup.PublicId)},
+			grantScopes:  []string{globals.GrantScopeThis, globals.GrantScopeDescendants},
+			roleScope:    globals.GlobalPrefix,
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   false,
+			},
+		},
+		{
+			name: "global_role_grant_all_specific_permissions",
+			grantStrings: []string{
+				fmt.Sprintf("ids=%s;types=group;actions=read", orgGroup.PublicId),
+				fmt.Sprintf("ids=%s;types=group;actions=read", projGroup.PublicId)},
+			grantScopes: []string{org.PublicId, proj.PublicId},
+			roleScope:   globals.GlobalPrefix,
+			getIdFound: map[string]bool{
+				globalGroup.PublicId: false,
+				orgGroup.PublicId:    true,
+				projGroup.PublicId:   true,
+			},
+		},
+	}
+
+	for _, tc := range testcases {
+		t.Run(tc.name, func(t *testing.T) {
+			// this creates everything required to get a token and creates context with auth token
+			acct := password.TestAccount(t, conn, authMethod.GetPublicId(), uuid.NewString())
+			user := iam.TestUser(t, iamRepo, globals.GlobalPrefix, iam.WithAccountIds(acct.GetPublicId()))
+			role := iam.TestRoleWithGrants(t, conn, tc.roleScope, tc.grantScopes, tc.grantStrings)
+			_ = iam.TestUserRole(t, conn, role.PublicId, user.PublicId)
+			fullGrantToken, err := atRepo.CreateAuthToken(ctx, user, acct.GetPublicId())
+			require.NoError(t, err)
+			fullGrantAuthCtx := auth.NewVerifierContext(requests.NewRequestContext(ctx, requests.WithUserId(user.GetPublicId())),
+				repoFn, atRepoFn, serversRepoFn, kmsCache, &authpb.RequestInfo{
+					PublicId:    fullGrantToken.PublicId,
+					Token:       fullGrantToken.GetToken(),
+					TokenFormat: uint32(auth.AuthTokenTypeBearer),
+				})
+			for id, found := range tc.getIdFound {
+				_, err := s.GetGroup(fullGrantAuthCtx, &pbs.GetGroupRequest{
+					Id: id,
+				})
+				// not found means expect error
+				if !found {
+					require.Error(t, err)
+					continue
 				}
-				require.ElementsMatch(t, tc.wantIDs, gotIDs)
-			})
-		}
-	})
+				require.NoError(t, err)
+
+			}
+		})
+	}
+
 }
